requirements.txt
python-telegram-bot[job-queue]
openai
python-dotenv
Flask
requests